SomeProtocols/dns_enumeration/dns_digger/requirements.txt
dnspython==2.8.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
rich==14.1.0
